+++ src/commands/mod/warn.ts
@@ -3,7 +3,7 @@ import { Message, MessageEmbed, TextChannel } from 'discord.js'
 import GuildConfig from '../../interfaces/GuildConfig';
 
 module.exports.run = async (msg: Message, args: string[], guildConf: GuildConfig) => {
-    if (!msg.member.hasPermission('KICK_MEMBERS')) return;
+    if (!msg.member.hasPermission('MANAGE_MESSAGES')) return;
 
     if (args.length >= 1 && msg.channel.type !== 'dm') {
         if (msg.mentions.everyone) return;
@@ -13,6 +13,8 @@ module.exports.run = async (msg: Message, args: string[], guildConf: GuildConfig
 
         if (!mention) return;
         if (mention.id === msg.author.id || mention.id === Kwako.user.id) return;
+
+        if (msg.author.id !== msg.guild.ownerID && mention.hasPermission('MANAGE_MESSAGES')) return;
 
         try {
             await msg.delete();
